Add osa 4d 4.21 answer

+++ controllers/blogs.js
@@ -45,9 +45,25 @@ blogsRouter.post("/", async (request, response, next) => {
 	}
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-	await BlogPost.findByIdAndRemove(request.params.id);
-	response.status(204).end();
+blogsRouter.delete("/:id", async (request, response, next) => {
+	try {
+		const decodedToken = jwt.verify(request.token, process.env.SECRET);
+		if (!request.token || !decodedToken.id) {
+			return response
+				.status(401)
+				.json({ error: "token missing or invalid" });
+		}
+
+		const user = await User.findById(decodedToken.id);
+		const blog = await BlogPost.findById(request.params.id);
+
+		if (user.id.toString() === blog.user.toString()) {
+			await BlogPost.findByIdAndRemove(request.params.id);
+			response.status(200).end();
+		}
+	} catch (exception) {
+		next(exception);
+	}
 });
 
 blogsRouter.put("/:id", async (request, response) => {
